requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
curl_cffi==0.12.0
distro==1.9.0
frozendict==2.4.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
lxml==6.0.0
MarkupSafe==3.0.2
multitasking==0.0.12
numpy==2.3.1
openai==1.98.0
pandas==2.3.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
plumbum==1.9.0
ply==3.11
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
requests==2.32.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yfinance==0.2.65
